Refactor assertions into trait
Use new RuntimeTrait in tests

wip

fix tests
