+++ code/parachain/frame/fnft/src/test/nonfungibles/transfer.rs
@@ -1,9 +1,7 @@
 use std::collections::{BTreeMap, BTreeSet};
 
 use codec::Encode;
-use composable_tests_helpers::test::{
-	block::process_and_progress_blocks, helper::assert_last_event,
-};
+use composable_tests_helpers::test::{block::process_and_progress_blocks, helper::RuntimeTrait};
 use composable_traits::{
 	account_proxy::{AccountProxy, ProxyType},
 	fnft::FinancialNft,
@@ -83,7 +81,7 @@ fn roundtrip() {
 
 		// send one of ALICE's NFTs to BOB
 		assert_ok!(Nft::transfer(&TEST_COLLECTION_ID, &nft_to_trade, &BOB));
-		assert_last_event::<MockRuntime>(Event::Nft(crate::Event::FinancialNftTransferred {
+		MockRuntime::assert_last_event(Event::Nft(crate::Event::FinancialNftTransferred {
 			collection_id: TEST_COLLECTION_ID,
 			instance_id: nft_to_trade,
 			to: BOB,
@@ -96,7 +94,7 @@ fn roundtrip() {
 
 		// send said NFT back
 		assert_ok!(Nft::transfer(&TEST_COLLECTION_ID, &nft_to_trade, &ALICE));
-		assert_last_event::<MockRuntime>(Event::Nft(crate::Event::FinancialNftTransferred {
+		MockRuntime::assert_last_event(Event::Nft(crate::Event::FinancialNftTransferred {
 			collection_id: TEST_COLLECTION_ID,
 			instance_id: nft_to_trade,
 			to: ALICE,
@@ -147,7 +145,7 @@ fn many() {
 
 		// transfer one of ALICE's NFTs to BOB
 		assert_ok!(Nft::transfer(&TEST_COLLECTION_ID, &a0, &BOB));
-		assert_last_event::<MockRuntime>(Event::Nft(crate::Event::FinancialNftTransferred {
+		MockRuntime::assert_last_event(Event::Nft(crate::Event::FinancialNftTransferred {
 			collection_id: TEST_COLLECTION_ID,
 			instance_id: a0,
 			to: BOB,
@@ -170,7 +168,7 @@ fn many() {
 		// transfer all of CHARLIES's NFTs to BOB
 		for nft_id in charlies_nfts.iter() {
 			assert_ok!(Nft::transfer(&TEST_COLLECTION_ID, nft_id, &BOB));
-			assert_last_event::<MockRuntime>(Event::Nft(crate::Event::FinancialNftTransferred {
+			MockRuntime::assert_last_event(Event::Nft(crate::Event::FinancialNftTransferred {
 				collection_id: TEST_COLLECTION_ID,
 				instance_id: *nft_id,
 				to: BOB,
@@ -194,7 +192,7 @@ fn many() {
 
 		// transfer one of (what was originally CHARLIES's) NFTs from BOB to ALICE
 		assert_ok!(Nft::transfer(&TEST_COLLECTION_ID, &c9, &ALICE));
-		assert_last_event::<MockRuntime>(Event::Nft(crate::Event::FinancialNftTransferred {
+		MockRuntime::assert_last_event(Event::Nft(crate::Event::FinancialNftTransferred {
 			collection_id: TEST_COLLECTION_ID,
 			instance_id: c9,
 			to: ALICE,
@@ -216,7 +214,7 @@ fn many() {
 
 		// transfer one of (what was originally CHARLIES's) NFTs from ALICE back to CHARLIE
 		assert_ok!(Nft::transfer(&TEST_COLLECTION_ID, &c9, &CHARLIE),);
-		assert_last_event::<MockRuntime>(Event::Nft(crate::Event::FinancialNftTransferred {
+		MockRuntime::assert_last_event(Event::Nft(crate::Event::FinancialNftTransferred {
 			collection_id: TEST_COLLECTION_ID,
 			instance_id: c9,
 			to: CHARLIE,
