+++ code/parachain/frame/staking-rewards/Cargo.toml
@@ -32,8 +32,6 @@ sp-std = { default-features = false, git = "https://github.com/paritytech/substr
 [dev-dependencies]
 env_logger = "0.9.0"
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.27" }
-hex-literal = "0.3.3"
-once_cell = "1.8.0"
 orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library", rev = "377213f750755cc48e80a3131eaae63b5eda8362" }
 orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", rev = "377213f750755cc48e80a3131eaae63b5eda8362", default-features = false }
 pallet-account-proxy = { path = "../../frame/account-proxy", default-features = false }
