+++ code/parachain/frame/staking-rewards/src/test/test_update_reward_pools.rs
@@ -1,24 +1,23 @@
 use composable_tests_helpers::test::{
 	block::process_and_progress_blocks,
 	currency::{PICA, USDT, XPICA},
-	helper::assert_extrinsic_event,
+	helper::RuntimeTrait,
 };
 use composable_traits::staking::{
 	Reward, RewardConfig, RewardPoolConfiguration, RewardRate, RewardRatePeriod, RewardUpdate,
 };
 use frame_support::{traits::TryCollect, BoundedBTreeMap};
 
-use super::prelude::{MINIMUM_STAKING_AMOUNT, STAKING_FNFT_COLLECTION_ID};
-
-use crate::test::{
-	default_lock_config, mint_assets, new_test_ext,
-	prelude::{
-		add_to_rewards_pot_and_assert, block_seconds, create_rewards_pool_and_assert,
-		ONE_YEAR_OF_BLOCKS,
+use crate::{
+	runtime::{MaxRewardConfigsPerPool, Origin, StakingRewards, Test, ALICE},
+	test::{
+		default_lock_config, mint_assets, new_test_ext,
+		prelude::{
+			block_seconds, MINIMUM_STAKING_AMOUNT, ONE_YEAR_OF_BLOCKS, STAKING_FNFT_COLLECTION_ID,
+		},
+		test_reward_accumulation_hook::{check_rewards, CheckRewards, PoolRewards},
 	},
-	runtime::{self, MaxRewardConfigsPerPool, Origin, StakingRewards, ALICE},
-	test_reward_accumulation_hook::{check_rewards, CheckRewards, PoolRewards},
-	Test,
+	test_helpers::{add_to_rewards_pot_and_assert, create_rewards_pool_and_assert},
 };
 
 #[test]
@@ -31,31 +30,29 @@ fn test_update_reward_pool() {
 		const INITIAL_REWARD_RATE_AMOUNT: u128 = 10;
 		const UPDATED_REWARD_RATE_AMOUNT: u128 = 5;
 
-		create_rewards_pool_and_assert::<Test, runtime::Event>(
-			RewardPoolConfiguration::RewardRateBasedIncentive {
-				owner: ALICE,
-				asset_id: PICA::ID,
-				start_block: 2,
-				end_block: ONE_YEAR_OF_BLOCKS + 1,
-				reward_configs: [(
-					USDT::ID,
-					RewardConfig {
-						max_rewards: 1_000_u128,
-						reward_rate: RewardRate::per_second(INITIAL_REWARD_RATE_AMOUNT),
-					},
-				)]
-				.into_iter()
-				.try_collect()
-				.expect("Rewards pool has a valid config for creation; QED"),
-				lock: default_lock_config(),
-				share_asset_id: XPICA::ID,
-				financial_nft_asset_id: STAKING_FNFT_COLLECTION_ID,
-				minimum_staking_amount: MINIMUM_STAKING_AMOUNT,
-			},
-		);
+		create_rewards_pool_and_assert::<Test>(RewardPoolConfiguration::RewardRateBasedIncentive {
+			owner: ALICE,
+			asset_id: PICA::ID,
+			start_block: 2,
+			end_block: ONE_YEAR_OF_BLOCKS + 1,
+			reward_configs: [(
+				USDT::ID,
+				RewardConfig {
+					max_rewards: 1_000_u128,
+					reward_rate: RewardRate::per_second(INITIAL_REWARD_RATE_AMOUNT),
+				},
+			)]
+			.into_iter()
+			.try_collect()
+			.expect("Rewards pool has a valid config for creation; QED"),
+			lock: default_lock_config(),
+			share_asset_id: XPICA::ID,
+			financial_nft_asset_id: STAKING_FNFT_COLLECTION_ID,
+			minimum_staking_amount: MINIMUM_STAKING_AMOUNT,
+		});
 
 		mint_assets([ALICE], [USDT::ID], INITIAL_AMOUNT);
-		add_to_rewards_pot_and_assert(ALICE, PICA::ID, USDT::ID, INITIAL_AMOUNT);
+		add_to_rewards_pot_and_assert::<Test>(ALICE, PICA::ID, USDT::ID, INITIAL_AMOUNT);
 
 		process_and_progress_blocks::<StakingRewards, Test>(1);
 
@@ -79,7 +76,7 @@ fn test_update_reward_pool() {
 		.try_collect()
 		.unwrap();
 
-		assert_extrinsic_event::<Test, _, _, _, _>(
+		Test::assert_extrinsic_event(
 			StakingRewards::update_rewards_pool(Origin::root(), PICA::ID, reward_updates),
 			crate::Event::RewardPoolUpdated { pool_id: PICA::ID },
 		);
